refactored rendering of courses and lessons into components to bbe mapped, fixed console warnings
